requirements.txt
requests==2.26.0
python-dotenv==0.19.1
PyYAML==6.0
Flask==2.0.1
Werkzeug==2.0.1
flask-socketio==5.1.1